src/requirements.txt
about-time==4.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
alive-progress==3.2.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
awsipranges==0.3.3
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.21.0
azure-keyvault-keys==4.10.0
azure-mgmt-applicationinsights==4.1.0
azure-mgmt-authorization==4.0.0
azure-mgmt-compute==34.0.0
azure-mgmt-containerregistry==12.0.0
azure-mgmt-containerservice==34.1.0
azure-mgmt-core==1.5.0
azure-mgmt-cosmosdb==9.7.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-monitor==6.0.2
azure-mgmt-network==28.1.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.3.0
azure-mgmt-search==9.1.0
azure-mgmt-security==7.0.0
azure-mgmt-sql==3.0.1
azure-mgmt-storage==22.1.1
azure-mgmt-subscription==3.1.1
azure-mgmt-web==8.0.0
azure-storage-blob==12.24.1
blinker==1.9.0
boto3==1.35.99
botocore==1.35.99
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
colorama==0.4.6
cryptography==44.0.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.18
detect-secrets==1.5.0
Django==5.2.1
djangorestframework==3.16.0
dnspython==2.7.0
durationpy==0.10
email_validator==2.2.0
filelock==3.18.0
Flask==3.0.3
frozenlist==1.6.0
google-api-core==2.25.0rc1
google-api-python-client==2.163.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grapheme==0.6.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
MarkupSafe==3.0.2
microsoft-kiota-abstractions==1.9.2
microsoft-kiota-authentication-azure==1.9.2
microsoft-kiota-http==1.9.2
microsoft-kiota-serialization-form==1.9.2
microsoft-kiota-serialization-json==1.9.2
microsoft-kiota-serialization-multipart==1.9.2
microsoft-kiota-serialization-text==1.9.2
msal==1.32.3
msal-extensions==1.3.1
msgraph-core==1.3.3
msgraph-sdk==1.23.0
msrest==0.7.1
multidict==6.4.4
narwhals==1.41.0
nest-asyncio==1.6.0
numpy==2.0.2
oauthlib==3.2.2
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
pandas==2.2.3
plotly==6.1.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
prowler==5.7.1
py-ocsf-models==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.21
PyGithub==2.5.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
retrying==1.3.4
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.10.4
schema==0.7.7
shodan==1.31.0
six==1.17.0
slack_sdk==3.34.0
sniffio==1.3.1
sqlparse==0.5.3
std-uritemplate==2.0.5
tabulate==0.9.0
tldextract==5.3.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.0.6
wrapt==1.17.2
XlsxWriter==3.2.3
yarl==1.20.0
zipp==3.22.0
